Fix handling of methods+classes
We no longer treat methods separately. They are just functions, and
should be treated as such. The only difference now is when we invoke
a function on an instance that had to be queried from a parent class,
we will shift the arguments by 1 and prepend `this`. It is slower than
using the hack where we use slot 0, but this complicates handling
dynamic situations where you may want to add a method to a class at
runtime.